requirements.txt
tkinter 
pillow